Added headers to ProcessStateChange.hh and moved Register struct to new Register.hh.
